+++ astropy/io/misc/asdf/extension.py
@@ -7,20 +7,23 @@ import os
 # the extension class so that _astropy_asdf_types is populated correctly. We
 # could do this using __init__ files, except it causes pytest import errors in
 # the case that asdf is not installed.
-from astropy.io.asdf.tags.fits.fits import *
-from astropy.io.asdf.tags.table.table import *
-from astropy.io.asdf.tags.time.time import *
-from astropy.io.asdf.tags.transform.basic import *
-from astropy.io.asdf.tags.transform.compound import *
-from astropy.io.asdf.tags.transform.polynomial import *
-from astropy.io.asdf.tags.transform.projections import *
-from astropy.io.asdf.tags.transform.tabular import *
-from astropy.io.asdf.tags.unit.quantity import *
-from astropy.io.asdf.tags.unit.unit import *
-from astropy.io.asdf.types import _astropy_asdf_types
+from astropy.io.misc.asdf.tags.fits.fits import *
+from astropy.io.misc.asdf.tags.table.table import *
+from astropy.io.misc.asdf.tags.time.time import *
+from astropy.io.misc.asdf.tags.transform.basic import *
+from astropy.io.misc.asdf.tags.transform.compound import *
+from astropy.io.misc.asdf.tags.transform.polynomial import *
+from astropy.io.misc.asdf.tags.transform.projections import *
+from astropy.io.misc.asdf.tags.transform.tabular import *
+from astropy.io.misc.asdf.tags.unit.quantity import *
+from astropy.io.misc.asdf.tags.unit.unit import *
+from astropy.io.misc.asdf.types import _astropy_asdf_types
 
 from asdf.extension import BuiltinExtension
 from asdf.resolver import Resolver, DEFAULT_URL_MAPPING
+
+
+__all__ = ['AstropyAsdfExtension']
 
 
 # This extension is used to register custom tag types that have schemas defined
