+++ astropy/io/misc/asdf/tags/fits/setup_package.py
@@ -3,5 +3,5 @@
 
 def get_package_data():  # pragma: no cover
     return {
-        str('astropy.io.asdf.tags.fits.tests'): ['data/*.fits']
+        str('astropy.io.misc.asdf.tags.fits.tests'): ['data/*.fits']
     }
