+++ astropy/io/misc/asdf/tags/table/table.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 from astropy import table
-from astropy.io.asdf.types import AstropyAsdfType
+from astropy.io.misc.asdf.types import AstropyAsdfType
 
 from asdf import yamlutil
 from asdf.tags.core.ndarray import NDArrayType
