+++ astropy/io/misc/asdf/tags/unit/unit.py
@@ -4,7 +4,7 @@
 import six
 
 from astropy.units import Unit, UnitBase
-from astropy.io.asdf.types import AstropyAsdfType
+from astropy.io.misc.asdf.types import AstropyAsdfType
 
 
 class UnitType(AstropyAsdfType):
